Move search component with the buttons
At the beginning of our application the search bar is locate it in the
center of the browser. When the user clicks the "Buscar" button, all the
search section needs to move at the top of the browser.
So, the Search section will help us to handle this with a bool flag
called "isAtTop" and a method called "handleOpenCloseSearchBar" that will
change the flag's value. This method we are going to pass by props to
the SearchBox component that is located the buttons that will handle the
behavior of the section.