Use React context API
In this commit I worked in how call API and understand how the Context API
works

The first thing that I did was create a method to call the Pokemon API
with an object that will be the arguments to do the request.

To use the Context API we require two things, the main Context (we could
have one or many) that will be our main data in the application, and we
needa Provider that will pass all the data to our components that need
our data. The `children` props that is in the `PokemonProvider` means
that will wrap the component that will use the data