Render all pokemons
With the components PokemonList and PokemonListItem will render all the
components. The PokemonList is going render the pokemon and call the PokemonListItem that
has the template of how we will display a pokemon.

The PokemonListItem use the Link feature of react-router because if the
user click the button to see more details about the pokemon, we have to
pass the pokemon id and for that the method getId() is going to split
the url with the purpose of get just the id