Create PokemonDetail component
At this point of time, we subscribe to the provider to get two things.
The first thing is the method to call the API and we pass the pokemon id
to have all its details.
The second thing is the pokemonDetail that will store the details when
we called the API